fix(components): fix as comment
