+++ .github/workflows/jest.yml
@@ -11,10 +11,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - uses: pnpm/action-setup@v4
+      - uses: pnpm/action-setup@v2
         name: Install pnpm
         with:
-          version: 8.14.1
+          version: 8
           run_install: false
       - name: Run Tests
         run: |
